Move names and descriptions
