+++ Chemo/Chemo.csproj
@@ -1,5 +1,6 @@
 ﻿<?xml version="1.0" encoding="utf-8"?>
-<Project ToolsVersion="4.0" DefaultTargets="Build" xmlns="http://schemas.microsoft.com/developer/msbuild/2003">
+<Project ToolsVersion="12.0" DefaultTargets="Build" xmlns="http://schemas.microsoft.com/developer/msbuild/2003">
+  <Import Project="..\packages\ILRepack.2.0.16\build\ILRepack.props" Condition="Exists('..\packages\ILRepack.2.0.16\build\ILRepack.props')" />
   <PropertyGroup>
     <Configuration Condition=" '$(Configuration)' == '' ">Debug</Configuration>
     <Platform Condition=" '$(Platform)' == '' ">AnyCPU</Platform>
@@ -7,9 +8,12 @@
     <OutputType>WinExe</OutputType>
     <NoStandardLibraries>false</NoStandardLibraries>
     <AssemblyName>Chemo</AssemblyName>
-    <TargetFrameworkVersion>v4.5</TargetFrameworkVersion>
+    <TargetFrameworkVersion>v4.6</TargetFrameworkVersion>
     <MinimumVisualStudioVersion>15</MinimumVisualStudioVersion>
     <TargetPlatformVersion>10</TargetPlatformVersion>
+    <NuGetPackageImportStamp>
+    </NuGetPackageImportStamp>
+    <TargetFrameworkProfile />
   </PropertyGroup>
   <PropertyGroup>
     <RootNamespace>Chemo</RootNamespace>
@@ -39,6 +43,7 @@
     <Reference Include="Windows.Storage" />
   </ItemGroup>
   <ItemGroup>
+    <None Include="app.config" />
     <None Include="Chemo.manifest">
       <SubType>Designer</SubType>
     </None>
@@ -63,24 +68,24 @@
       <DependentUpon>Resources.resx</DependentUpon>
     </Compile>
     <Compile Include="RegistryUtils.cs" />
-    <Compile Include="Treatment\DeprovisionStoreApps.cs" />
-    <Compile Include="Treatment\DisableCortana.cs" />
+    <Compile Include="Treatment\Apps\DeprovisionStoreApps.cs" />
+    <Compile Include="Treatment\Apps\DisableCortana.cs" />
     <Compile Include="frmAbout.cs">
       <SubType>Form</SubType>
     </Compile>
     <Compile Include="frmAbout.Designer.cs">
       <DependentUpon>frmAbout.cs</DependentUpon>
     </Compile>
-    <Compile Include="Treatment\DisableInternetSearchResults.cs" />
-    <Compile Include="Treatment\GameBar.cs" />
-    <Compile Include="Treatment\InternetExplorer.cs" />
+    <Compile Include="Treatment\Config\DisableInternetSearchResults.cs" />
+    <Compile Include="Treatment\Config\GameBar.cs" />
+    <Compile Include="Treatment\Features\InternetExplorer.cs" />
     <Compile Include="Treatment\BaseTreatment.cs" />
-    <Compile Include="Treatment\OneDrive.cs" />
-    <Compile Include="Treatment\RemoveStoreApps.cs" />
-    <Compile Include="Treatment\RequireCtrlAltDel.cs" />
-    <Compile Include="Treatment\SetClockUTC.cs" />
-    <Compile Include="Treatment\SuggestedApps.cs" />
-    <Compile Include="Treatment\WindowsUpdateReboot.cs" />
+    <Compile Include="Treatment\Apps\OneDrive.cs" />
+    <Compile Include="Treatment\Apps\RemoveStoreApps.cs" />
+    <Compile Include="Treatment\Config\RequireCtrlAltDel.cs" />
+    <Compile Include="Treatment\Config\SetClockUTC.cs" />
+    <Compile Include="Treatment\Config\SuggestedApps.cs" />
+    <Compile Include="Treatment\Config\WindowsUpdateReboot.cs" />
     <Compile Include="ChemoTreeView.cs">
       <SubType>Component</SubType>
     </Compile>
@@ -152,4 +157,10 @@
   <ProjectExtensions>
     <VisualStudio AllowExistingFolder="true" />
   </ProjectExtensions>
+  <Target Name="EnsureNuGetPackageBuildImports" BeforeTargets="PrepareForBuild">
+    <PropertyGroup>
+      <ErrorText>This project references NuGet package(s) that are missing on this computer. Use NuGet Package Restore to download them.  For more information, see http://go.microsoft.com/fwlink/?LinkID=322105. The missing file is {0}.</ErrorText>
+    </PropertyGroup>
+    <Error Condition="!Exists('..\packages\ILRepack.2.0.16\build\ILRepack.props')" Text="$([System.String]::Format('$(ErrorText)', '..\packages\ILRepack.2.0.16\build\ILRepack.props'))" />
+  </Target>
 </Project>
