+++ Chemo/Treatment/BaseTreatment.cs
@@ -4,11 +4,20 @@ namespace Chemo.Treatment
     {
         private readonly Logger logger = Logger.Instance;
 
-        public static string Name { get; }
-        public static string Description { get; }
+        /// <summary>
+        /// The short name of the treatment.
+        /// </summary>
+        /// <returns>The treatment name.</returns>
+        public abstract string Name();
 
         /// <summary>
-        /// Determines whether this treatment should be applied. This operation should be idempotent.
+        /// The tooltip text of the treatment.
+        /// </summary>
+        /// <returns>The treatment tooltip.</returns>
+        public abstract string Tooltip();
+
+        /// <summary>
+        /// Determines whether the treatment should be applied. This operation should be idempotent.
         /// </summary>
         /// <returns>Returns true if the treatment should be applied, false otherwise.</returns>
         public abstract bool ShouldPerformTreatment();
