+++ Chemo/Treatment/Config/GameBar.cs
@@ -1,7 +1,7 @@
 using Microsoft.Win32;
 using System;
 
-namespace Chemo.Treatment
+namespace Chemo.Treatment.Config
 {
     class GameBar : BaseTreatment
     {
@@ -9,6 +9,16 @@ namespace Chemo.Treatment
         private static readonly string GameDVR = @"HKEY_CURRENT_USER\SOFTWARE\Microsoft\Windows\CurrentVersion\GameDVR";
         private static readonly string GameConfigStore = @"HKEY_CURRENT_USER\System\GameConfigStore";
         private static readonly int DesiredValue = 0;
+
+        public override string Name()
+        {
+            return "Turn Off Game Bar";
+        }
+
+        public override string Tooltip()
+        {
+            return "Turns off the game bar for both apps and games.";
+        }
 
         public override bool ShouldPerformTreatment()
         {
