+++ Chemo/packages.config
@@ -1,4 +1,5 @@
-<?xml version="1.0" encoding="utf-8"?>
+﻿<?xml version="1.0" encoding="utf-8"?>
 <packages>
-  <package id="Microsoft.Dism" version="2.0.2" targetFramework="net45" />
+  <package id="ILRepack" version="2.0.16" targetFramework="net45" />
+  <package id="Microsoft.Dism" version="2.0.2" targetFramework="net45" developmentDependency="true" />
 </packages>
